Add .prettierrc.json
Much the same as I explained here:

https://github.com/liferay/liferay-npm-build-tools/pull/278#issuecomment-469221124

I've found that when I update the eslint and eslint-config-liferay
dependencies in this repo and then run "npm run format" it wants to make
about 5k lines of changes that you can see here:

https://gist.github.com/wincent/707b67c46b11fa0c21720ffbf84eaf4a

So, bring across a "standard" .prettierrc.json like the ones we have in
alloy-editor, clay, liferay-themes-sdk, and in that
liferay-npm-build-tools PR. As you can see, it produces only minimal
changes, like the ones seen in this commit.

Thinking about it, I'm guessing it might be because the move to the
latest version of eslint-config-liferay involves the removal of a number
of rules that would otherwise conflict with Prettier, so the output
format has changed just a little bit, as seen in this commit.